More complex struct field updates work
The compiler is now able to emit code to update fields in arbitrarily
nested structs. One limitation (or rather bug, if we want to be
completely consistent) is that field updates do not return a value.
But this is not a big problem - they are executed mostly for their side
effects anyway.