Fix: the main expression of a function's body is not a toplevel
Since it *must* return a value it is not in a "toplevel position". Every
expression that exists in a context where it does not have to return a
usable slot is in a "toplevel position". All expressions that exist in a
context where they *do* have to return a usable slot (e.g. the main
expression of a function body) are not in a "toplevel position".

The "toplevel position" is thus a badly chosen name, but let's fix it
later.