Fixes needed on Viua VM side
The code works... but the VM somehow becomes corrupted during deferred
call execution and segfaults. This needs to be investigated, but except
for this bug both Viuact and the VM are capable of running the
"impressive example" program.

Viuact would not need (yet) a more elaborate error handling mechanism
than process isolation, but the deferred call bug renders it unusable.
Therefore, a special-purpose error handling construct should be added to
Viuact. The syntax could be like this:

    (try <expression> with
        (<error-tag> <name> <handling-expression>)+
    )

The expression would evaluate to either the value of <expression> if no
error happened, or to a value of one of <handling-expression>s if an
error with matching <error-tag> was encountered. If the encountered
error matches no <error-tag> it is propagated up the stack.