Allow setting watchdog functions
Watchdog is a function that is triggered when an actor crashes. It
receives a "death message" containing information about the exception
that caused the crash and is the last line of defense we have against
going into uncontrolled failure mode.

Watchdog may be set to send a message to a controller actor, to notify
it that some worker actor died. That way the controller actor may just
passively wait for messages (death and other) instead of actively
maintaining some kind of "heartbeat" protocol.