Get hash of the main module
In the future the hash may be used to check if it needs to be
recompiled. Avoiding recompilation of modules that did not change may
speed the compilation substantially.

Currently, the solution to slow compilation speeds may be to avoid
nested modules and create programs as collections of small, disjoint
modules and use Make to track whether they need to be recompiled or not.