Support label punning
Imagine we have a function like this:

    (let fn (~a ~b) ...)

Assume that preparing values for parameters ~a and ~b is a process
slightly more involved than usual. So we create a pair of variables to
hold their values before actually calling the function:

    (let a { ... })
    (let b { ... })

It would be repetitive, cumbersome, and silly to have to write code like
this to pass them to the function:

    (fn (~a a) (~b b))

Let the compiler handle the repetition if we already took care to name
the variables appropriately. With label punning the code looks like
this:

    (fn ~a ~b)

Much better.