Add file I/O example
The program reads itself and writes the result to standard output.
Remember that I/O in Viuact works the same as in Viua VM, and that means
asynchronously, so you only get an I/O request out of Io.read and
Io.write calls and have to actually wait for the request to complete
using the Io.wait built-in function.