Add a new front command for other Viuact tools
Instead of

    $ viuact-cc --mode exec file.lisp

you can now use

    $ viuact cc --mode exec file.lisp

The fact that all commands are now under a single wrapper (a la Git) is
a little bit more elegant. The utility executable can now be moved to
another directory to avoid cluttering /whatever/bin.

Also, the new front has a built-in "switch" command which can be used to
create and manage semi-isolated environments for Viua VM and Viuact. The
idea came from OCaml's "opam switch".

First create a new environment named "foo":

    $ viuact switch make foo

This will create a necessary directory structre (by default in the
~/.local/viuact/switch directory), and copy Viua VM execuatbles from
system directories there. In effect, this makes a snapshot of a Viua VM
executable set.

Then, activate the environment:

    $ `viuact switch to foo`

The backquotes are there to make your shell evaluate the exports
necessary to setup the environment.

You can also list the environments you have (the active one is marked
with a star), and remove them:

    $ viuact switch ls
     bar
    *foo
     baz
    $ viuact switch rm baz

Checking the current environment status is also simple:

    $ viuact switch status
    switch: test
    path:   /home/johndoe/.local/viuact/switch/here/test