Deallocate slots that are will not be used
We are able to prove statically that some slots will not be used after
some point. Use this information to deallocate them and allow the
compiler to reuse them.

With this simple optimisation, in some cases, the actual register
pressure goes down to about 70% of the static pressure (mostly in long
functions).

More savings can be made by tracking actual usage of registers, and by
changing ordering of instructions in call instructions.

Also, try to more aggressively detect unused values in unused arms of
if-expressions, match-expressions, etc.