Allow marking slots as "dropping" or explicitly void
Such a slot would be considered void even if it has an index assigned.
This makes it possible to create "dummy" slots for expressions to put
their value into, and make the value disappear, which is incredibly
useful for expressions whose value is unused (their values can be put
into such dummy slots).