Unify type of the enum with type of the value used in constructor
Since we know that the 'a in the enum field constructor's is, for
example, i8 there is no reason not to unify these two types. This way
the compiler knows that the enum is:

    ((i8) example)

and not just:

    (('a) example)

With this knowledge it can provide additional error checks net when
analysing the code.