WIP: Parse more types
There are two kinds of val declaration:

- value: (val name type)
- function: (val name (type...) -> type)

Of course, type can be:

- simple: i8
- polymorphic: (('a) option)
- function: (() -> i8)

And they can be freely combined, which sometimes gets unwieldy. For
example:

    (val ('a) map (
        (('a) option)
        (((('a) option)) -> (('a) option))
    ) -> (('a) option))

Declares type of function named "map" which takes two parameters:
- the first one is an "option" enum parametrised by 'a
- the second one is a function which takes a single parameter: "option"
  enum parametrised by 'a, and returns "option" enum parametrised by 'a'
The "map" function returns an "option" enum parametrised by 'a.

Yeah... While the information is there the amount of parentheses creates
noise which hides it. Without some kind of type alises it will be
painful to specify some types.