Detect source root, main file, and what kind of module is compiled
Viuact compiler uses source hierarchy to determine the full name of a
compiled module. For example, the following directory structures:

    Std/POSIX/IO.vt
    Mod/Nest/Foo.vt

yield such modules:

    Std::POSIX::IO
    Mod::Nest::Foo

To facilitate good practices of putting source files in their own
directory Viauct compiler tries to determine "source root", ie, the
directory which is the base from which path-module name mapping should
begin. For example, given the following path:

    src/Mod/Foo.vt

the compiler infers that the source root is src/ and the "useful" path
is Mod/Foo.vt, so the module's name will be Mod::Foo.

In case the compiler infers the source root incorrectly, a -r option is
provided to override compiler's own decision. Assuming the source root
in the previous example was in fact src/Mod, the compiler should be
invoked like this:

    $ viuact cc -r src/Mod src/Mod/Foo.vt

The full path to the source file must be specified.