Merge v2 to the development branch, making it semi-official
The v2 code is stable enough if you don't try to hurt it too much, and
don't look at it funny. I guess that's enough for devel: people can use
it, but if it occasionally breaks... no harm, no foul, it's still an
early release, alphish-state software. You can't expect it to be
*rock-solid* at this stage, and I never promised that.

Anyway, v2 is now the semi-official main release of Viuact.

(More like 0.3.0 than v2, but I digress.)